Get rid of io_k8s_client_go dependency in pkg/apis/contrail/v1alpha1/ packege
Motivation behind this change is to reduce number of dependencies in this api package.
This package may be imported by external clients/users and they shouldn't be forced to import unnecessary packages
