+++ pkg/apis/contrail/v1alpha1/BUILD.bazel
@@ -13,7 +13,6 @@ go_library(
         "control_types.go",
         "defaults.go",
         "doc.go",
-        "exec_to_pod.go",
         "fernetkeymanager_types.go",
         "keystone_types.go",
         "kubemanager_types.go",
@@ -53,13 +52,8 @@ go_library(
         "@io_k8s_apimachinery//pkg/runtime:go_default_library",
         "@io_k8s_apimachinery//pkg/runtime/schema:go_default_library",
         "@io_k8s_apimachinery//pkg/types:go_default_library",
-        "@io_k8s_client_go//dynamic:go_default_library",
-        "@io_k8s_client_go//kubernetes:go_default_library",
-        "@io_k8s_client_go//rest:go_default_library",
-        "@io_k8s_client_go//tools/remotecommand:go_default_library",
         "@io_k8s_kube_openapi//pkg/common:go_default_library",
         "@io_k8s_sigs_controller_runtime//pkg/client:go_default_library",
-        "@io_k8s_sigs_controller_runtime//pkg/client/config:go_default_library",
         "@io_k8s_sigs_controller_runtime//pkg/controller/controllerutil:go_default_library",
         "@io_k8s_sigs_controller_runtime//pkg/reconcile:go_default_library",
         "@io_k8s_sigs_controller_runtime//pkg/runtime/scheme:go_default_library",
