+++ pkg/apis/contrail/v1alpha1/cassandra_types.go
@@ -209,11 +209,6 @@ func (c *Cassandra) UpdateSTS(sts *appsv1.StatefulSet, instanceType string, requ
 	return UpdateSTS(sts, instanceType, request, reconcileClient, strategy)
 }
 
-// PodIPListAndIPMapFromInstance gets a list with POD IPs and a map of POD names and IPs.
-func (c *Cassandra) PodIPListAndIPMapFromInstance(instanceType string, request reconcile.Request, reconcileClient client.Client) (*corev1.PodList, map[string]string, error) {
-	return PodIPListAndIPMapFromInstance(instanceType, &c.Spec.CommonConfiguration, request, reconcileClient, false, true, false, false, false, false)
-}
-
 //PodsCertSubjects gets list of Cassandra pods certificate subjets which can be passed to the certificate API
 func (c *Cassandra) PodsCertSubjects(podList *corev1.PodList, serviceIP string) []certificates.CertificateSubject {
 	altIPs := PodAlternativeIPs{ServiceIP: serviceIP}
