+++ pkg/apis/contrail/v1alpha1/rabbitmq_types.go
@@ -309,11 +309,6 @@ func (c *Rabbitmq) UpdateSTS(sts *appsv1.StatefulSet, instanceType string, reque
 	return UpdateSTS(sts, instanceType, request, reconcileClient, strategy)
 }
 
-// PodIPListAndIPMapFromInstance gets a list with POD IPs and a map of POD names and IPs.
-func (c *Rabbitmq) PodIPListAndIPMapFromInstance(instanceType string, request reconcile.Request, reconcileClient client.Client) (*corev1.PodList, map[string]string, error) {
-	return PodIPListAndIPMapFromInstance(instanceType, &c.Spec.CommonConfiguration, request, reconcileClient, true, false, false, false, false, false)
-}
-
 //PodsCertSubjects gets list of Rabbitmq pods certificate subjets which can be passed to the certificate API
 func (c *Rabbitmq) PodsCertSubjects(podList *corev1.PodList) []certificates.CertificateSubject {
 	var altIPs PodAlternativeIPs
