+++ pkg/apis/contrail/v1alpha1/vrouter_types.go
@@ -295,11 +295,6 @@ func (c *Vrouter) SetInstanceActive(client client.Client, activeStatus *bool, ds
 	return nil
 }
 
-// PodIPListAndIPMapFromInstance gets a list with POD IPs and a map of POD names and IPs.
-func (c *Vrouter) PodIPListAndIPMapFromInstance(instanceType string, request reconcile.Request, reconcileClient client.Client, getPhysicalInterface bool, getPhysicalInterfaceMac bool, getPrefixLength bool, getGateway bool) (*corev1.PodList, map[string]string, error) {
-	return PodIPListAndIPMapFromInstance(instanceType, &c.Spec.CommonConfiguration, request, reconcileClient, false, true, getPhysicalInterface, getPhysicalInterfaceMac, getPrefixLength, getGateway)
-}
-
 //PodsCertSubjects gets list of Vrouter pods certificate subjets which can be passed to the certificate API
 func (c *Vrouter) PodsCertSubjects(podList *corev1.PodList) []certificates.CertificateSubject {
 	var altIPs PodAlternativeIPs
