+++ pkg/apis/contrail/v1alpha1/webui_types.go
@@ -294,11 +294,6 @@ func (c *Webui) UpdateSTS(sts *appsv1.StatefulSet, instanceType string, request 
 	return UpdateSTS(sts, instanceType, request, reconcileClient, strategy)
 }
 
-// PodIPListAndIPMapFromInstance gets a list with POD IPs and a map of POD names and IPs.
-func (c *Webui) PodIPListAndIPMapFromInstance(instanceType string, request reconcile.Request, reconcileClient client.Client) (*corev1.PodList, map[string]string, error) {
-	return PodIPListAndIPMapFromInstance(instanceType, &c.Spec.CommonConfiguration, request, reconcileClient, true, false, false, false, false, false)
-}
-
 //PodsCertSubjects gets list of Config pods certificate subjets which can be passed to the certificate API
 func (c *Webui) PodsCertSubjects(podList *corev1.PodList) []certificates.CertificateSubject {
 	var altIPs PodAlternativeIPs
