+++ pkg/controller/provisionmanager/BUILD.bazel
@@ -11,8 +11,10 @@ go_library(
     deps = [
         "//pkg/apis/contrail/v1alpha1:go_default_library",
         "//pkg/certificates:go_default_library",
+        "//pkg/configuration:go_default_library",
         "//pkg/controller/utils:go_default_library",
         "@com_github_ghodss//:go_default_library",
+        "@in_gopkg_yaml.v2//:go_default_library",
         "@io_k8s_api//apps/v1:go_default_library",
         "@io_k8s_api//core/v1:go_default_library",
         "@io_k8s_apimachinery//pkg/api/errors:go_default_library",
