+++ pkg/controller/utils/utils.go
@@ -1,14 +1,25 @@
 package utils
 
 import (
+	"context"
+	"errors"
+	"fmt"
+	"net"
+	"strings"
+
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/event"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
+	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	"github.com/Juniper/contrail-operator/pkg/apis/contrail/v1alpha1"
+	"github.com/Juniper/contrail-operator/pkg/k8s"
 )
 
 var log = logf.Log.WithName("utils")
@@ -530,4 +541,171 @@ func GetContainerFromList(containerName string, containerList []*v1alpha1.Contai
 		}
 	}
 	return nil
+}
+
+// PodIPListAndIPMapFromInstance gets a list with POD IPs and a map of POD names and IPs.
+func PodIPListAndIPMapFromInstance(instanceType string,
+	commonConfiguration *v1alpha1.PodConfiguration,
+	request reconcile.Request,
+	reconcileClient client.Client,
+	waitForInit bool,
+	getHostname bool,
+	getInterface bool,
+	getMac bool,
+	getPrefix bool,
+	getGateway bool) (*corev1.PodList, map[string]string, error) {
+	var podNameIPMap = make(map[string]string)
+	labelSelector := labels.SelectorFromSet(map[string]string{"contrail_manager": instanceType,
+		instanceType: request.Name})
+	listOps := &client.ListOptions{Namespace: request.Namespace, LabelSelector: labelSelector}
+	podList := &corev1.PodList{}
+	err := reconcileClient.List(context.TODO(), podList, listOps)
+	if err != nil {
+		return &corev1.PodList{}, map[string]string{}, err
+	}
+	if len(podList.Items) > 0 {
+		if waitForInit {
+			if int32(len(podList.Items)) == *commonConfiguration.Replicas {
+				podNameIPMap, err = getPodInitStatus(reconcileClient, podList, getHostname, getInterface, getMac, getPrefix, getGateway, waitForInit)
+				if err != nil {
+					return podList, podNameIPMap, err
+				}
+			}
+			if int32(len(podNameIPMap)) != *commonConfiguration.Replicas {
+				return &corev1.PodList{}, map[string]string{}, nil
+			}
+		} else if len(podList.Items) > 0 {
+			podNameIPMap, err = getPodInitStatus(reconcileClient, podList, getHostname, getInterface, getMac, getPrefix, getGateway, waitForInit)
+			if err != nil {
+				return podList, podNameIPMap, err
+			}
+		}
+		return podList, podNameIPMap, nil
+	}
+	return &corev1.PodList{}, map[string]string{}, nil
+}
+
+func GetPodsHostname(c client.Reader, pod *corev1.Pod) (string, error) {
+	if !pod.Spec.HostNetwork {
+		return pod.Spec.Hostname, nil
+	}
+	n := corev1.Node{}
+	if err := c.Get(context.Background(), types.NamespacedName{Name: pod.Spec.NodeName}, &n); err != nil {
+		return "", err
+	}
+
+	for _, a := range n.Status.Addresses {
+		if a.Type == corev1.NodeHostName {
+			return a.Address, nil
+		}
+	}
+
+	return "", errors.New("couldn't get pods hostname")
+}
+
+func getPodInitStatus(reconcileClient client.Client,
+	podList *corev1.PodList,
+	getHostname bool,
+	getInterface bool,
+	getMac bool,
+	getPrefix bool,
+	getGateway bool,
+	waitForInit bool) (map[string]string, error) {
+	var podNameIPMap = make(map[string]string)
+	for idx, pod := range podList.Items {
+		if !waitForInit && pod.Status.Phase != "Running" && pod.Status.Phase != "Pending" {
+			return map[string]string{}, nil
+		}
+		if pod.Status.PodIP == "" {
+			return map[string]string{}, nil
+		}
+		if getHostname || getInterface || getMac || getPrefix {
+			for _, initStatus := range pod.Status.InitContainerStatuses {
+				if initStatus.Name == "init" {
+					if initStatus.State.Terminated == nil {
+						if initStatus.State.Running != nil {
+							annotationMap := pod.GetAnnotations()
+							if annotationMap == nil {
+								annotationMap = make(map[string]string)
+							}
+							if getHostname {
+
+								hostname, err := GetPodsHostname(reconcileClient, &pod)
+								if err != nil {
+									return map[string]string{}, err
+								}
+								annotationMap["hostname"] = hostname
+							}
+							if getInterface {
+								command := []string{"/bin/sh", "-c", "ifconfig | sed -n '/addr:" + pod.Status.PodIP + "/{g;h;p};h;x'  |awk '{print $1}'"}
+								physicalInterface, _, err := k8s.ExecToPodThroughAPI(command, "init", pod.Name, pod.Namespace, nil)
+								if err != nil {
+									return map[string]string{}, fmt.Errorf("failed getting interface")
+								}
+								annotationMap["physicalInterface"] = strings.Trim(physicalInterface, "\n")
+							}
+							if getGateway {
+								command := []string{"/bin/sh", "-c", "ip route get 1.1.1.1 |grep -v cache |sed -e 's/.* via \\(.*\\) dev.*/\\1/'"}
+								gateway, _, err := k8s.ExecToPodThroughAPI(command, "init", pod.Name, pod.Namespace, nil)
+								if err != nil {
+									return map[string]string{}, fmt.Errorf("failed getting gateway")
+								}
+								annotationMap["gateway"] = strings.Trim(gateway, "\n")
+							}
+							if getMac {
+								command := []string{"/bin/sh", "-c", "ip address show | sed -n '/inet " + pod.Status.PodIP + "\\//{g;h;p};h;x' |awk '{print $2}'"}
+								physicalInterfaceMac, _, err := k8s.ExecToPodThroughAPI(command, "init", pod.Name, pod.Namespace, nil)
+								if err != nil {
+									return map[string]string{}, fmt.Errorf("failed getting mac")
+								}
+								annotationMap["physicalInterfaceMac"] = strings.Trim(physicalInterfaceMac, "\n")
+							}
+							if getPrefix {
+								command := []string{"/bin/sh", "-c", "ip addr sh |sed -n 's/.*" + pod.Status.PodIP + "\\/\\([^ ]*\\).*/\\1/p'"}
+								prefixLength, _, err := k8s.ExecToPodThroughAPI(command, "init", pod.Name, pod.Namespace, nil)
+								if err != nil {
+									return map[string]string{}, fmt.Errorf("failed getting prefix")
+								}
+								annotationMap["prefixLength"] = strings.Trim(prefixLength, "\n")
+							}
+
+							if cidr, ok := pod.Annotations["dataSubnet"]; ok {
+								if cidr != "" {
+									command := []string{"/bin/sh", "-c", "ip r | grep " + cidr + " | awk -F' ' '{print $NF}'"}
+									addr, _, err := k8s.ExecToPodThroughAPI(command, "init", pod.Name, pod.Namespace, nil)
+									if err != nil {
+										return map[string]string{}, fmt.Errorf("failed getting ip address from data subnet")
+									}
+									ip := strings.Trim(addr, "\n")
+									if net.ParseIP(ip) != nil {
+										annotationMap["dataSubnetIP"] = ip
+									} else {
+										return map[string]string{}, fmt.Errorf("no valid ip from data subnet")
+									}
+								}
+							}
+
+							podList.Items[idx].SetAnnotations(annotationMap)
+							(&podList.Items[idx]).SetAnnotations(annotationMap)
+							foundPod := &corev1.Pod{}
+							err := reconcileClient.Get(context.TODO(), types.NamespacedName{Name: pod.Name, Namespace: pod.Namespace}, foundPod)
+							if err != nil {
+								return map[string]string{}, err
+							}
+							foundPod.SetAnnotations(annotationMap)
+							err = reconcileClient.Update(context.TODO(), foundPod)
+							if err != nil {
+								return map[string]string{}, err
+							}
+							podList.Items[idx] = *foundPod
+						} else {
+							return map[string]string{}, nil
+						}
+					}
+				}
+			}
+		}
+		podNameIPMap[pod.Name] = pod.Status.PodIP
+	}
+	return podNameIPMap, nil
 }
