+++ pkg/k8s/BUILD.bazel
@@ -5,6 +5,7 @@ go_library(
     srcs = [
         "cluster_info.go",
         "config_map.go",
+        "exec.go",
         "k8s.go",
         "owner.go",
         "secret.go",
@@ -21,9 +22,12 @@ go_library(
         "@io_k8s_apimachinery//pkg/runtime:go_default_library",
         "@io_k8s_apimachinery//pkg/runtime/schema:go_default_library",
         "@io_k8s_apimachinery//pkg/types:go_default_library",
+        "@io_k8s_client_go//kubernetes:go_default_library",
         "@io_k8s_client_go//kubernetes/typed/core/v1:go_default_library",
+        "@io_k8s_client_go//tools/remotecommand:go_default_library",
         "@io_k8s_sigs_controller_runtime//pkg/client:go_default_library",
         "@io_k8s_sigs_controller_runtime//pkg/client/apiutil:go_default_library",
+        "@io_k8s_sigs_controller_runtime//pkg/client/config:go_default_library",
         "@io_k8s_sigs_controller_runtime//pkg/controller/controllerutil:go_default_library",
     ],
 )
