+++ test/e2e/aio/BUILD.bazel
@@ -27,6 +27,7 @@ go_test(
         "//pkg/client/kubeproxy:go_default_library",
         "//pkg/client/swift:go_default_library",
         "//pkg/controller/utils:go_default_library",
+        "//pkg/k8s:go_default_library",
         "//test/env/client:go_default_library",
         "//test/logger:go_default_library",
         "//test/wait:go_default_library",
